Update PhabricatorEmailLoginController.php
The existing password reset message is terrible advice, in what I think was just an attempt at sarcasm. Unfortunately sarcasm is often lost on many users and this messaging will at best end up being confusing and create more work for teams that support phabricator.

Ideally, this messaging would be configurable by admins, but in the absence of that feature, I'm proposing new language that keeps the spirit of the original but with better security advice without any sarcasm.