Revert "Restrict all crumbs to 240 max width" for this release
Summary:
Ref T12287. See D17361. That fixed a mostly-theoretical bug with crumbs named things like "MMMMMMMMMMMMMMMMMMMM", but caused a less-theoretical buggy side effect in Safari.

For now, just keep the "MMMMMMMM" crumbs around since that's the easiest/least-bad/safest fix prior to the release cut. We can fix this more broadly when we have more time to look at it.

Test Plan: Looked at profiles, saw the entire name for "hector" and the too-long-crumb for "MMMMM".

Reviewers: chad

Reviewed By: chad

Maniphest Tasks: T12287

Differential Revision: https://secure.phabricator.com/D17382